iio: accel: sca3300: modified to support multi chips
The drive support sca3300 only.
There are some other similar chips, for instance, SCL3300.
Prepare the way for multiple chips and additional channels.
Modify the driver to read the device id.
Add the tables for the corresponding id to support multiple chips.
Add prepares for the addition of extra channels.
Add prepares for handling the operation modes for multiple chips.

Reviewed-by: Rob Herring <robh@kernel.org>
Reviewed-by: Andy Shevchenko <andy.shevchenko@gmail.com>
Reviewed-by: Jonathan Cameron <jic23@kernel.org>
Signed-off-by: LI Qingwu <Qing-wu.Li@leica-geosystems.com.cn>