iio: accel: sca3300: modified to support multi chips
The driver supports sca3300 only, there are some other similar chips,
for instance, SCL3300. This commit prepares the way for multiple chips
and additional channels. Modify the driver to read the device ID and load
the corresponding sensor information from the table to support multiple
chips. add prepares for the addition of extra channels. Add prepares for
handling the operation modes for multiple chips.

Signed-off-by: LI Qingwu <Qing-wu.Li@leica-geosystems.com.cn>