iio: accel: bmi088: Modified the scale calculate.
Modified the accelerator scale calculation
from the fixed formula to a scale list.
The scales in the list are calculated by 1/sensitivity*9.8.
The reason for modifying the scales is:
The units after application of scale are not m/s^2,
due to 980 being used instead of 9.8 in the current driver,
The formula is only for the device with the range of 3, 6, 12, 24g,
but there are some other chips has a range of 2, 4, 6, 8g.

Signed-off-by: LI Qingwu <Qing-wu.Li@leica-geosystems.com.cn>