media: uapi: add nal unit header fields to encode_params
The VEPU540 and VEPU580 may prepare the NAL unit headers in the coded
params. These values have to be provided by user space and be written
into the hardware registers.

Furthermore, nal_ref_idc indicates if a picture will be used as
reference and is a hint to the driver, if it needs to keep the
reconstructed buffer or not.

Upstream-Status: Pending

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>