drm/bridge: fsl-ldb: add 2nd bridge support
Register a 2nd bridge if the 2nd ldb channel is connected to a panel
node as well which differs from the frist panel node. This allows the
CRTC to register two encoder devices later on.

Since the DRM uAPI is based on encoders, not connectors two encoders are
necessary to independently control two lvds panels which are connected
via an external lvds mux:

                                    +--> LVDS-panel-1
                                    |
    ctrc --> fsl-ldb --> LVDS-mux --+
                                    |
                                    +--> LVDS-panel-2

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>